+++ src/engine/sidechain/shoutconnection.h
@@ -43,11 +43,11 @@ class ShoutConnection
     void write(const unsigned char* header, const unsigned char* body,
                int headerLen, int bodyLen) override;
     // gets stream position
-    int tell() override;
+    qint64 tell() override;
     // sets stream position
-    void seek(int pos) override;
+    void seek(qint64 pos) override;
     // gets stream length
-    int filelen() override;
+    qint64 filelen() override;
 
     /** connects to server **/
     bool serverConnect();
